pyproject.toml
[project]
name = "musicplane"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "numpy>=2.3.3",
    "pillow>=11.3.0",
    "pyfluidsynth>=1.3.4",
    "pygame>=2.6.1",
    "pyganim>=0.9.2",
    "sounddevice>=0.5.2",
]
